% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{file_exists}
\alias{file_exists}
\title{Checks if a file exists}
\usage{
file_exists(file, wd = getwd())
}
\arguments{
\item{wd}{A Path to the file, if not set otherwise the working directory is used.}

\item{x}{A filename.}
}
\value{
TRUE if file exists, FALSE if not.
}
\description{
Checks if files exists after converting relative into absolute path based on 
current working directory by default.
}
\examples{
file <- system.file("inst/template.docx", package = "R2wdaddons")
file_exists(file)
}
\keyword{internal}
