% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2wdaddons-pkg.R
\docType{package}
\name{R2wdaddons}
\alias{R2wdaddons}
\alias{R2wdaddons-package}
\title{Addon to the R2wd package.}
\description{
Creates and modifies MS Word documents with R.
Builds on the R2wd package and adds some extra functionality.
}

