% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacements.R
\name{to_win_path}
\alias{to_win_path}
\title{Convert path slashes to path with double backslashes.}
\usage{
to_win_path(file)
}
\arguments{
\item{file}{A file path.}
}
\value{
Path with double backslashes instead of slahes.
}
\description{
RDCOMClient / the COM interface seems to only support this type of path.
}
