% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addons.R
\name{wdAddImageCaption}
\alias{wdAddImageCaption}
\title{Adds a caption to an image.}
\usage{
wdAddImageCaption(i, ishp = NULL, title = "", sep = ":", wdapp = .R2wd)
}
\arguments{
\item{i}{The number of the image inside the document.}

\item{ishp}{A pointer to the image. Best used with wdReplaceTextByImage return value.}

\item{title}{The title of the image.}
}
\description{
This function adds a caption to an image.
}
\examples{
wdGet("inst/wdScaleImage.docx")
wdAddImageCaption(i = 1, title = "My image")
}

