% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replacements.R
\name{wdGet2}
\alias{wdGet2}
\title{wdGet2}
\usage{
wdGet2(filename = NULL, path = getwd(), method = "RDCOMClient",
  visible = TRUE)
}
\arguments{
\item{path}{A Path to the file, if not set otherwise the working directory is used.}

\item{method}{Use \code{"rcom"} (default) or \code{"RDCOMClient"}.}

\item{visible}{If the Applciation should be visible.}

\item{x}{A filename.}
}
\value{
Store invisible file handle named in global environment.
}
\description{
There is a bug in \code{R2wd::wdGet} as path slashes are not replaced by double backslashes.
This may cause an error when using RCDOMClient on some plattforms. This version simply
does the replacement.
}
\examples{
wdGet2("inst/test.docx")

# save as temp file
file <- tempfile(fileext=".docx")
wdSave2(file)

# Fails: Does not yet notice that the path is absolute
# wdGet2(file)

}

