% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addons.R
\name{wdReplaceTextByImage}
\alias{wdReplaceTextByImage}
\title{Replaces a string with an Image}
\usage{
wdReplaceTextByImage(find = "", file, wdapp = .R2wd, wd = getwd(),
  warn = TRUE)
}
\arguments{
\item{find}{A string.}

\item{file}{path to the image.}
}
\value{
invisibly returns pointer to new inlineshape object.
}
\description{
This function replaces a piece of text with an image.
}
\examples{
wdGet("inst/test.docx")
p1 <- wdReplaceTextByImage("text", "inst/image.png")
}

