% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{wdReplaceTextByTable}
\alias{wdReplaceTextByTable}
\title{Replaces a string with a table created from a dataframe}
\usage{
wdReplaceTextByTable(find, x, wdapp = .R2wd)
}
\arguments{
\item{find}{A string.}

\item{x}{A dataframe.}
}
\value{
Invisibly returns a pointer to new table object.
}
\description{
This function replaces a piece of text with a table created using a dataframe.
}
\section{TODO}{
 handle case if text is found multiple times. May a version createTable which can be 
passed a Range object from the search function is te better choice.s
}

\examples{
file <- system.file("inst/template.docx", package = "R2wdaddons")
wdGet2(file)

# add a table at placeholder
d <- mtcars[1:10, 1:6]
tbl <- wdReplaceTextByTable("[placeholder 3]", d)

}
