% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacements.R
\name{wdSave2}
\alias{wdSave2}
\title{wdSave2}
\usage{
wdSave2(file = NULL, path = getwd(), wdapp = .R2wd)
}
\arguments{
\item{file}{File name (if missing, Word will ask).}

\item{path}{Path to file. Current working directory is used as default.}

\item{wdapp}{The handle to the Word Application (usually not needed).}
}
\description{
There is a bug in \code{R2wd::wdSave} as path slashes are not replaced by double backslashes.
This may cause an error when using RCDOMClient on some plattforms. This version simply 
does the replacement.
}
\examples{
wdGet2("inst/test.docx")

# save as temp file
file <- tempfile(fileext=".docx")
wdSave2(file)

# Fails: Does not yet notice that the path is absolute
# wdGet2(file)

}
