% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addons.R
\name{wdScaleImage}
\alias{wdScaleImage}
\title{Changes the scale of an image}
\usage{
wdScaleImage(i, ishp = NULL, width = NULL, height = NULL,
  lock.asp = TRUE, units = "px", wdapp = .R2wd)
}
\arguments{
\item{i}{The number of the image inside the document.}

\item{ishp}{A pointer to the image. Best used with wdReplaceTextByImage return value.}

\item{width}{The width of the new image in Percent of the word page.}

\item{height}{The height of the new image in Percent of the word page.}

\item{lock.asp}{Locks the ratio of hight and width (correct this way?)}

\item{units}{The number of pixel the image should have}
}
\value{
invisibly returns pointer to new inlineshape object.
}
\description{
This function change the sclaing of an image in a document.
}
\examples{
wdGet("inst/wdScaleImage.docx")
wdScaleImage(1, width=70, units="w")
}

