% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addons.R
\name{wdSearchString}
\alias{wdSearchString}
\title{Checks if a sting in contained in a word document.}
\usage{
wdSearchString(find = "", wdapp = .R2wd, warn = TRUE)
}
\arguments{
\item{find}{A string.}
}
\value{
TRUE if string is found, FALSE if not.
}
\description{
This function checks if a sting in contained in a word document.
}
\examples{
wdGet("inst/test.docx")
wdSearchString("Some text")
}

