% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addons.R
\name{file_exists}
\alias{file_exists}
\title{Checks if a file exists!}
\usage{
file_exists(file, wd = getwd())
}
\arguments{
\item{wd}{A Path to the file, if not set otherwise the working directory is used.}

\item{x}{A filename.}
}
\value{
TRUE if file exists, FALSE if not.
}
\description{
This function checks if a file is existing in a directory!
}
\examples{
file_exists("inst/test.docx")
}
\keyword{internal}

