% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{wdAddImageCaption}
\alias{wdAddImageCaption}
\title{Adds a caption to an image.}
\usage{
wdAddImageCaption(i, ishp = NULL, title = "", sep = ":", label = -1,
  position = 1, wdapp = .R2wd)
}
\arguments{
\item{i}{The number of the image inside the document.}

\item{ishp}{A pointer to the image. Best used with wdReplaceTextByImage return value.}

\item{title}{The title of the image.}

\item{label}{WdCaptionLabelID enumeration. -1 = Figure caption (default). 
You may use as string for self customized captions.}

\item{position}{0=above, 1=below (default).}
}
\description{
This function adds a caption to an image.
}
\section{TODO}{
 check if captions exists and replace it.
}

\examples{
file <- system.file("inst/template.docx", package = "R2wdaddons")
wdGet2(file)
wdAddImageCaption(i = 1, title = "My image")
}
