% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{wdReplaceTextByImage}
\alias{wdReplaceTextByImage}
\title{Replaces a string with an image}
\usage{
wdReplaceTextByImage(find = "", file, wdapp = .R2wd, wd = getwd(),
  warn = TRUE)
}
\arguments{
\item{find}{A string.}

\item{file}{The path to the image.}
}
\value{
Invisibly returns a pointer to new inlineshape object.
}
\description{
This function replaces a piece of text with an image.
}
\section{TODO}{
 handle case if text is found multiple times
}

\examples{
wdGet("inst/test.docx")
p1 <- wdReplaceTextByImage("text", "inst/image.png")
}
