% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{wdReplaceTextByImage}
\alias{wdReplaceTextByImage}
\title{Replaces a string with an image}
\usage{
wdReplaceTextByImage(find = "", file, wdapp = .R2wd, wd = getwd(),
  warn = TRUE)
}
\arguments{
\item{find}{A string.}

\item{file}{The path to the image.}
}
\value{
Invisibly returns a pointer to new inlineshape object.
}
\description{
This function replaces a piece of text with an image.
}
\section{TODO}{
 handle case if text is found multiple times
}

\examples{
file <- system.file("inst/template.docx", package = "R2wdaddons")
wdGet2(file)

# image file to put in document
img <- system.file("inst/sky.jpeg", package = "R2wdaddons")

# replace and rescale first image
shp <- wdReplaceTextByImage("[placeholder 1]", img)
wdScaleImage(ishp = shp, width=200, units="px")
wdAddImageCaption(ishp = shp, title = "Image is 200 pixels wide")
}
