% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addons.R
\name{wdSaveAsPdf}
\alias{wdSaveAsPdf}
\title{Save active document as PDF file}
\usage{
wdSaveAsPdf(file = NULL, path = getwd(), open = FALSE, wdapp = .R2wd)
}
\arguments{
\item{file}{File name (if missing, current name is used).}

\item{path}{Path to file. Current working directory is used as default.}

\item{open}{Open PDF file after coversion? Default is \code{FALSE}.}

\item{wdapp}{The handle to the Word Application (usually not needed).}
}
\description{
Save active document as PDF file
}
\examples{
wdGet2("inst/test.docx")          # open word doc in package

file <- tempfile(fileext=".pdf")  # file in temp directory
wdSaveAsPdf(file)
}

