% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{wdScaleImage}
\alias{wdScaleImage}
\title{Changes the scale of an image}
\usage{
wdScaleImage(i, ishp = NULL, width = NULL, height = NULL,
  lock.asp = TRUE, units = "px", wdapp = .R2wd)
}
\arguments{
\item{i}{The number of the image inside the document.}

\item{ishp}{A pointer to the inline shape, as e.g. returned by \code{\link{wdReplaceTextByImage}}.}

\item{width}{The width of the new image in Percent of the word page.}

\item{height}{The height of the new image in Percent of the word page.}

\item{lock.asp}{Locks the ratio of height and width (aspect ratio).}

\item{units}{The number of pixel the image should have}
}
\value{
invisibly returns pointer to new inlineshape object.
}
\description{
This function change the sclaing of an image in a document.
}
\section{TODO}{
 add example with wdReplaceTextByImage
}

\examples{
file <- system.file("inst/template.docx", package = "R2wdaddons")
wdGet2(file)

wdScaleImage(i = 1, width=70, units="w")
}
