% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{wdSearchString}
\alias{wdSearchString}
\title{Checks if a string is contained in a Word document.}
\usage{
wdSearchString(find = "", wdapp = .R2wd, warn = TRUE)
}
\arguments{
\item{find}{A string.}
}
\value{
\code{TRUE} if string is found, \code{FALSE} if not.
}
\description{
This function checks if a string in contained in a Word document.
}
\examples{
file <- system.file("inst/template.docx", package = "R2wdaddons")
wdGet2(file)

# does the search string exist?
wdSearchString("[placeholder 1]")
}
