% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{wdTocUpdate}
\alias{wdTocUpdate}
\title{Update table of contents}
\usage{
wdTocUpdate(wdapp = .R2wd)
}
\arguments{
\item{wdapp}{The handle to Word Application (usually not needed).}
}
\description{
Update table of contents
}
